requirements.txt
lmdb
gevent
gevent-websocket
websocket-client
